requirements.txt
Flask==2.2.2
pdf2image==1.16.0
pytesseract==0.3.10
requests==2.28.1
python-dotenv==1.0.0
Werkzeug==2.2.3
pdfplumber==0.5.28
opencv-python-headless==4.8.0.76
numpy<2